requirements.txt
# Flask Web Framework & Server
Flask
gunicorn
itsdangerous

# Document & Image Processing
Pillow
PyPDF2
pdf2image
pdfplumber
python-docx
reportlab

# OCR & Computer Vision
easyocr
numpy
opencv-python
pytesseract
torch
torchvision

# Email & Environment Variables
python-dotenv
sendgrid